Handle ticket clean up correctly with the default registry To properly remove tickets for all registry types.
